Refactored code for creating HnItem instances
We implemented a new trait `OptionalHnItem` which allowsobjects with optional title, description and guid fields. We added an implementation to `rss::Item` and changed the item creation process in the main function to use it. We also refactored HnItem creation: instead of using a `new` function which takes String parameters, we now have a `from_optional_item` function that takes a reference to a `OptionalHnItem` object, as well as a private `new` function which takes slice references. This simplified the test cases and reduced the amount of item creation code in the main function.